+++ metrics-core/src/main/java/com/yammer/metrics/reporting/ConsoleReporter.java
@@ -3,7 +3,7 @@ package com.yammer.metrics.reporting;
 import com.yammer.metrics.Metrics;
 import com.yammer.metrics.core.*;
 import com.yammer.metrics.stats.Snapshot;
-import com.yammer.metrics.util.MetricPredicate;
+import com.yammer.metrics.core.MetricPredicate;
 
 import java.io.PrintStream;
 import java.text.DateFormat;