+++ metrics-graphite/src/test/java/com/yammer/metrics/reporting/GraphiteReporterTest.java
@@ -1,9 +1,9 @@
 package com.yammer.metrics.reporting;
 
 import com.yammer.metrics.core.Clock;
+import com.yammer.metrics.core.MetricPredicate;
 import com.yammer.metrics.core.MetricsRegistry;
 import com.yammer.metrics.reporting.tests.AbstractPollingReporterTest;
-import com.yammer.metrics.util.MetricPredicate;
 
 import java.io.OutputStream;
 import java.net.Socket;